requirements.txt
# Dropbox & HTTP
dropbox>=11.0.0
requests>=2.28.0

# Data
pandas>=1.5.0
numpy>=1.23.0
pyarrow>=12.0.0

# Parsing
beautifulsoup4>=4.11.0

# LLM / Agents (for 13F table extraction fallback)
openai-agents>=0.1.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Gemini (optional; for LLM_PARSER_MODEL=gemini)
google-generativeai>=0.8.0
